LayoutTests/http/tests/webgpu/webgpu/api/validation/layout_shader_compat.html


                <script>
                    if (window.testRunner)
                        testRunner.waitUntilDone();
                    window.location.href = "/webgpu/cts.https.html?q=webgpu:api,validation,layout_shader_compat:*";
                </script>
            
